12basic 10undefined et curqui test better

+++ ft_format_u.c
@@ -6,7 +6,7 @@
 /*   By: dberger <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/27 18:11:54 by dberger           #+#    #+#             */
-/*   Updated: 2019/05/29 16:08:45 by dberger          ###   ########.fr       */
+/*   Updated: 2019/05/29 17:41:12 by dberger          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,21 +21,21 @@ t_printf		ft_width_u(t_printf save, int *j, unsigned long type)
 	w = save.width;
 	s = ft_sizenb_u(type);
 	count = w - s;
-	if (save.precision && w > save.precision && s < save.precision)
+	if (save.precision && w > save.precision && (int)s < save.precision)
 		count = w - save.precision;
-	if (save.precision && s < save.precision && w <= save.precision)
+	if (save.precision && (int)s < save.precision && w <= save.precision)
 		return (save);
 	if (save.flags & F_PLUS)
 		count--;
 	if ((save.flags & F_ZERO) && !(save.precision) && !(save.flags & F_MINUS))
 	{
-		while (s < w && (count--) > 0)
+		while ((int)s < w && (count--) > 0)
 			save = ft_check_add(save, j, '0');
 		return (save);
 	}
 	if (type == 0 && !(save.precision) && save.flags & F_POINT)
 		save = ft_check_add(save, j, ' ');
-	while (s < w && (count--) > 0)
+	while ((int)s < w && (count--) > 0)
 		save = ft_check_add(save, j, ' ');
 	return (save);
 }
@@ -47,7 +47,7 @@ t_printf		ft_precision_u(t_printf save, int *j, unsigned long type)
 
 	s = ft_sizenb_u(type);
 	p = save.precision;
-	if (s < p)
+	if ((int)s < p)
 	{
 		while (p - s > 0)
 		{
